aSpringFramework5-BeginnerToGuru/v-spring-mvc-content-negotiation/src/main/java/com/george/springframework/controllers/v1/VendorController.java
package com.george.springframework.controllers.v1;

import com.george.springframework.api.v1.model.VendorDTO;
import com.george.springframework.api.v1.model.VendorListDTO;
import com.george.springframework.services.VendorService;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

@Api(description = "This is my Vendor Controller")
@RestController
@RequestMapping(VendorController.BASE_URL)
public class VendorController {

    public static final String BASE_URL = "/api/v1/vendors/";

    private final VendorService vendorService;

    public VendorController(VendorService vendorService) {
        this.vendorService = vendorService;
    }

    @ApiOperation(value = "Get a list of vendors.", notes = "Those are some notes about the API.")
    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    public VendorListDTO getVendorList() {
        return vendorService.getAllVendors();
    }

    @ApiOperation(value = "Get a vendor by id.", notes = "Those are some notes about the API.")
    @GetMapping({"/id/{id}"})
    @ResponseStatus(HttpStatus.OK)
    public VendorDTO getVendorById(@PathVariable Long id) {
        return vendorService.getVendorById(id);
    }

    @ApiOperation(value = "Get a vendor by name.", notes = "Those are some notes about the API.")
    @GetMapping({"/name/{name}"})
    @ResponseStatus(HttpStatus.OK)
    public VendorDTO getVendorByName(@PathVariable String name) {
        return vendorService.getVendorByName(name);
    }

    @ApiOperation(value = "Create a new vendor.", notes = "Those are some notes about the API.")
    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public VendorDTO createNewVendor(@RequestBody VendorDTO vendorDTO){
        return vendorService.createNewVendor(vendorDTO);
    }

    @ApiOperation(value = "Update an existing vendor.", notes = "Those are some notes about the API.")
    @PutMapping({"/id/{id}"})
    @ResponseStatus(HttpStatus.OK)
    public VendorDTO updateVendor(@PathVariable Long id, @RequestBody VendorDTO vendorDTO){
        return vendorService.saveVendorByDTO(id, vendorDTO);
    }

    @ApiOperation(value = "Update a vendor property.", notes = "Those are some notes about the API.")
    @PatchMapping({"/id/{id}"})
    @ResponseStatus(HttpStatus.OK)
    public VendorDTO patchVendor(@PathVariable Long id, @RequestBody VendorDTO vendorDTO){
        return vendorService.saveVendorByDTO(id, vendorDTO);
    }

    @ApiOperation(value = "Delete a vendor.", notes = "Those are some notes about the API.")
    @DeleteMapping({"/id/{id}"})
    @ResponseStatus(HttpStatus.OK)
    public void deleteVendor(@PathVariable Long id){
        vendorService.deleteVendorById(id);
    }
}
